+++ CHANGELOG.md
@@ -15,7 +15,9 @@ Security - in case of vulnerabilities.
 
 ## [Unreleased]
 
-_TBD_
+Added
+
+- [x] Added new property to `Address`: `jobTitle`
 
 ## [2.15.0] 2022-02-25